bee_server/requirements.txt
flask
pymongo
google-cloud
google-cloud-storage